Cleaned code #64
Removed some unnecessary stuff, and minimalized react errors

+++ crypto-gaze/app/react-ui/src/Routes/ChartModifier.jsx
@@ -1,13 +1,9 @@
 import React, { useState, useEffect } from "react";
 import axios from 'axios';
 import { FormControl, InputLabel, Select, MenuItem } from "@mui/material";
-import MyChartSettings from "../Basic Components/MyChartSettings";
-import MyChart2Settings from "../Basic Components/MyChart2Settings";
 import { useNavigate } from "react-router-dom";
-import MyChartSettingsUpdate from "../Basic Components/MyChartSettingsUpdate";
-import MyChart2SettingsUpdate from "../Basic Components/MyChartSettings2Update";
 import ChartEditor from "../Basic Components/ChartEditor";
-import { Button, Card, FlexBox, Title, Text } from "@ui5/webcomponents-react";
+import { Button, Card, FlexBox, Title } from "@ui5/webcomponents-react";
 
 export function ChartModifier() {
 
@@ -16,7 +12,6 @@ export function ChartModifier() {
     const [chartSelected, setChartSelected] = useState();
     const [charts, setCharts] = useState([])
     const [chartData, setChartData] = useState([])
-    const [update, setUpdate] = useState("True")
     const [toggleEditor, setToggleEditor] = useState()
 
     const fetchCharts = async () => {
@@ -30,7 +25,7 @@ export function ChartModifier() {
         if (charts.length < 1) {
             fetchCharts();
         }
-        if (chartSelected != "") {
+        if (chartSelected !== "") {
             selectChart()
         }
         setToggleEditor(false)
@@ -39,28 +34,28 @@ export function ChartModifier() {
 
 
     const selectChart = () => {
-        console.log("Select Chart")
-        let chart = charts.find(item => item["id"] == chartSelected)
+        let chart = charts.find(item => item["id"] === chartSelected)
         setChartData(chart)
-        console.log(chart)
     }
 
 
     const updateChart = async () => {
-        const res = await axios.patch('http://localhost:4004/chart/CustomCharts/' + chartSelected, chartData, {
-            headers: {
-                "Authorization": "Basic admin",
-                "Content-Type": "application/json;IEEE754Compatible=true"
-            }
-        })
-        console.log(res)
-        navigate('/charts')
+        if (chartSelected !== "" && chartSelected !== "undefined" && chartSelected !== undefined) {
+            await axios.patch('http://localhost:4004/chart/CustomCharts/' + chartSelected, chartData, {
+                headers: {
+                    "Authorization": "Basic admin",
+                    "Content-Type": "application/json;IEEE754Compatible=true"
+                }
+            })
+            navigate('/charts')
+        }
     }
 
     const deleteChart = async () => {
-        const res = await axios.delete('http://localhost:4004/chart/CustomCharts/' + chartSelected)
-        console.log(res)
-        navigate('/charts')
+        if (chartSelected !== "" && chartSelected !== "undefined" && chartSelected !== undefined) {
+            await axios.delete('http://localhost:4004/chart/CustomCharts/' + chartSelected)
+            navigate('/charts')
+        }
     }
 
     const updaterFunction = (values) => {
@@ -69,19 +64,22 @@ export function ChartModifier() {
 
 
     const toggle = () => {
-        if (toggleEditor) {
-            setToggleEditor(false);
+        if (chartSelected !== "" && chartSelected !== "undefined" && chartSelected !== undefined) {
+            if (toggleEditor) {
+                setToggleEditor(false);
+            }
+            else {
+                setToggleEditor(true);
+            }
         }
-        else {
-            setToggleEditor(true);
-        }
+
     }
 
     return (
         <>
             <Card>
                 <FlexBox alignItems="Center" justifyContent="SpaceAround">
-                <Title>Select chart: </Title>
+                    <Title>Select chart: </Title>
                     <FormControl variant="outlined" sx={{ m: 1, minWidth: 200 }}>
                         <InputLabel>Chart title</InputLabel>
                         <Select
